ai-caller/pyproject.toml
[tool.poetry]
name = "ai-caller"
version = "0.1.0"
description = "A program which calls the api to generate an image and returns the image url"
authors = ["iharuya <i4haruya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
replicate = "^0.4.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
